requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pytz==2023.3
python-dotenv==1.0.0
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
blinker==1.6.3

# Development dependencies (optional)
pytest==7.4.2
pytest-flask==1.2.0
black==23.9.1
flake8==6.1.0

# Production dependencies (optional)
gunicorn==21.2.0
redis==5.0.1
celery==5.3.1
